requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
exceptiongroup==1.2.0
fake-headers==1.0.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
idna==3.4
lxml==4.9.3
magic-filter==1.0.9
multidict==6.0.4
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3
requests==2.30.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.13
tomli==2.0.1
transliterate==1.10.2
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.5.0
Unidecode==1.3.7
union==0.1.10
urllib3==2.0.2
webdriver-manager==4.0.1
webencodings==0.5.1
wsproto==1.2.0
yarl==1.9.2
